+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 All notable changes to this project will be documented in this file.
 # Changelog
 
+## [0.0.9] - 2019-08-21
+### Fixed
+- Method how menu class is import
+- Added proper Prefix & Suffix to menu classes
+- Some keymaps not show in Preference panel
+- Clearing keymaps
+
+## [0.0.8] - 2019-02-13
+### Fixed
+- Issue with quick render menu. Now render window will show like as dropdown menu (mentioned by Jacques Lucke)
+
 ## [0.0.7] - 2019-01-10
 ### Changed
 - Fixed typo in workspace Texture Paint > was TEXTURE_PAINTING
@@ -41,6 +52,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.9]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.9
+[0.0.8]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.8
 [0.0.7]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.7
 [0.0.6]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.6
 [0.0.5]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.5